Fix long casts / labels

+++ tlatools/org.lamport.tlatools/src/tlc2/TLC.java
@@ -653,7 +653,7 @@ public class TLC {
                 {
                     try
                     {
-                        TLCGlobals.chkptDuration = Integer.parseInt(args[index]) * 1000 * 60;
+                        TLCGlobals.chkptDuration = (long) Integer.parseInt(args[index]) * 1000 * 60;
                         if (TLCGlobals.chkptDuration < 0)
                         {
                             printErrorMsg("Error: expect a nonnegative integer for -checkpoint option.");
